# Exclude body contacts