--- a/newton/_src/sim/builder.py
+++ b/newton/_src/sim/builder.py
@@ -1350,7 +1350,11 @@ class ModelBuilder:
 
         if collision_filter_parent and parent > -1:
             for child_shape in self.body_shapes[child]:
+                if not self.shape_flags[child_shape] & ShapeFlags.COLLIDE_SHAPES:
+                    continue
                 for parent_shape in self.body_shapes[parent]:
+                    if not self.shape_flags[parent_shape] & ShapeFlags.COLLIDE_SHAPES:
+                        continue
                     # Ensure canonical order (smaller, larger) for consistent lookup
                     a, b = parent_shape, child_shape
                     if a > b:
@@ -1913,6 +1917,7 @@ class ModelBuilder:
         show_joint_keys=True,
         show_joint_types=True,
         plot_shapes=True,
+        show_shape_keys=True,
         show_shape_types=True,
         show_legend=True,
     ):
@@ -1926,6 +1931,7 @@ class ModelBuilder:
             show_joint_keys (bool): Whether to show the joint keys or indices
             show_joint_types (bool): Whether to show the joint types
             plot_shapes (bool): Whether to render the shapes connected to the rigid bodies
+            show_shape_keys (bool): Whether to show the shape keys or indices
             show_shape_types (bool): Whether to show the shape geometry types
             show_legend (bool): Whether to show a legend
         """
@@ -1976,10 +1982,12 @@ class ModelBuilder:
             vertices = ["-1"] + [str(i) for i in range(self.body_count)]
         if plot_shapes:
             for i in range(self.shape_count):
-                shape_label = f"shape_{i}"
+                shape_label = []
+                if show_shape_keys:
+                    shape_label.append(self.shape_key[i])
                 if show_shape_types:
-                    shape_label += f"\n({shape_type_str(self.shape_type[i])})"
-                vertices.append(shape_label)
+                    shape_label.append(f"\n({shape_type_str(self.shape_type[i])})")
+                vertices.append("\n".join(shape_label))
         edges = []
         edge_labels = []
         for i in range(self.joint_count):
@@ -2402,13 +2410,11 @@ class ModelBuilder:
             scale = (1.0, 1.0, 1.0)
         self.shape_body.append(body)
         shape = self.shape_count
-        if body in self.body_shapes:
+        if cfg.has_shape_collision:
             # no contacts between shapes of the same body
             for same_body_shape in self.body_shapes[body]:
                 self.shape_collision_filter_pairs.append((same_body_shape, shape))
-            self.body_shapes[body].append(shape)
-        else:
-            self.body_shapes[body] = [shape]
+        self.body_shapes[body].append(shape)
         self.shape_key.append(key or f"shape_{shape}")
         self.shape_transform.append(xform)
         self.shape_flags.append(cfg.flags)
@@ -2426,7 +2432,7 @@ class ModelBuilder:
         self.shape_collision_group.append(cfg.collision_group)
         self.shape_collision_radius.append(compute_shape_radius(type, scale, src))
         self.shape_group.append(self.current_env_group)
-        if cfg.collision_filter_parent and body > -1 and body in self.joint_parents:
+        if cfg.has_shape_collision and cfg.collision_filter_parent and body > -1 and body in self.joint_parents:
             for parent_body in self.joint_parents[body]:
                 if parent_body > -1:
                     for parent_shape in self.body_shapes[parent_body]: