--- a/newton/_src/solvers/mujoco/solver_mujoco.py
+++ b/newton/_src/solvers/mujoco/solver_mujoco.py
@@ -17,7 +17,6 @@ from __future__ import annotations
 
 import os
 import warnings
-from itertools import product
 from typing import TYPE_CHECKING, Any
 
 import numpy as np
@@ -1616,61 +1615,50 @@ class SolverMuJoCo(SolverBase):
             )
 
     @staticmethod
-    def find_body_collision_filter_pairs(colliding_shapes, shape_filters, shape_groups, shape_bodies, body_shapes):
+    def find_body_collision_filter_pairs(
+        model: Model,
+        selected_bodies: nparray,
+        colliding_shapes: nparray,
+    ):
         """For shape collision filter pairs, find body collision filter pairs that are contained within."""
-        colliding_shapes = set(colliding_shapes)
-        body_filters = set()
-        body_non_filterable = set()
-        remaining_shape_filters = []
-        for sp in shape_filters:
-            if sp[0] not in colliding_shapes or sp[1] not in colliding_shapes:
-                continue
-            b1, b2 = shape_bodies[sp[0]], shape_bodies[sp[1]]
-            bp = min(b1, b2), max(b1, b2)
-            if bp in body_filters:
-                continue
-            if bp in body_non_filterable:
-                remaining_shape_filters.append(sp)
-                continue
-            b1_shapes = {s for s in body_shapes[bp[0]] if s in colliding_shapes}
-            b2_shapes = {s for s in body_shapes[bp[1]] if s in colliding_shapes}
-            if any(
-                SolverMuJoCo.shapes_can_collide(bs1, bs2, shape_filters, shape_groups)
-                for bs1, bs2 in product(b1_shapes, b2_shapes)
-            ):
-                body_non_filterable.add(bp)
-                remaining_shape_filters.append(sp)
-            else:
-                body_filters.add(bp)
-        return list(body_filters), remaining_shape_filters
+
+        body_exclude_pairs = []
+        shape_set = set(colliding_shapes)
+
+        bodies_a, bodies_b = np.triu_indices(len(selected_bodies), k=1)
+        for body_a, body_b in zip(bodies_a, bodies_b, strict=True):
+            b1, b2 = selected_bodies[body_a], selected_bodies[body_b]
+            shapes_1 = model.body_shapes[b1]
+            shapes_2 = model.body_shapes[b2]
+            shapes_1 = [s for s in shapes_1 if s in shape_set]
+            shapes_2 = [s for s in shapes_2 if s in shape_set]
+            excluded = True
+            for shape_1 in shapes_1:
+                for shape_2 in shapes_2:
+                    if (shape_1, shape_2) not in model.shape_collision_filter_pairs:
+                        excluded = False
+                        break
+            if excluded:
+                body_exclude_pairs.append((b1, b2))
+        return body_exclude_pairs
 
     @staticmethod
-    def shapes_can_collide(a, b, shape_filter_pairs=None, shape_group=None):
-        if a == b:
-            return False
-        if shape_filter_pairs is not None and ((a, b) in shape_filter_pairs or (b, a) in shape_filter_pairs):
-            return False
-
-        # skip env groups
-
-        if shape_group is not None:
-            group_a, group_b = shape_group[a], shape_group[b]
-
-            # Mirrors warp funcs in broad_phase_common.py
-            if group_a == 0 or group_b == 0:
-                return False
-            if group_a > 0 and not (group_a == group_b or group_b < 0):
-                return False
-            elif group_a < 0 and not (group_a != group_b):
-                return False
-        return True
-
-    @staticmethod
-    def color_collision_shapes(model: Model, selected_shapes: nparray, visualize_graph: bool = False) -> np.ndarray:
+    def color_collision_shapes(
+        model: Model, selected_shapes: nparray, visualize_graph: bool = False, shape_keys: list[str] | None = None
+    ) -> nparray:
         """
         Find a graph coloring of the collision filter pairs in the model.
         Shapes within the same color cannot collide with each other.
         Shapes can only collide with shapes of different colors.
+
+        Args:
+            model (Model): The model to color the collision shapes of.
+            selected_shapes (nparray): The indices of the collision shapes to color.
+            visualize_graph (bool): Whether to visualize the graph coloring.
+            shape_keys (list[str]): The keys of the shapes, only used for visualization.
+
+        Returns:
+            nparray: An integer array of shape (num_shapes,), where each element is the color of the corresponding shape.
         """
         # we first create a mapping from selected shape to local color shape index
         # to reduce the number of nodes in the graph to only the number of selected shapes
@@ -1679,33 +1667,37 @@ class SolverMuJoCo(SolverBase):
         for i, shape in enumerate(selected_shapes):
             to_color_shape_index[shape] = i
         # find graph coloring of collision filter pairs
-        collision_group = model.shape_collision_group
+        num_shapes = len(selected_shapes)
+        shape_a, shape_b = np.triu_indices(num_shapes, k=1)
+        cgroup = [model.shape_collision_group[i] for i in selected_shapes]
         # edges representing colliding shape pairs
         graph_edges = [
-            (to_color_shape_index[i], to_color_shape_index[j])
-            for i, j in product(selected_shapes, selected_shapes)
-            if i != j
-            and (
-                ((i, j) not in model.shape_collision_filter_pairs and (j, i) not in model.shape_collision_filter_pairs)
-                or collision_group[i] != collision_group[j]
+            (i, j)
+            for i, j in zip(shape_a, shape_b, strict=True)
+            if (
+                (selected_shapes[i], selected_shapes[j]) not in model.shape_collision_filter_pairs
+                and (cgroup[i] == cgroup[j] or cgroup[i] == -1 or cgroup[j] == -1)
             )
         ]
+        shape_color = np.zeros(model.shape_count, dtype=np.int32)
         if len(graph_edges) > 0:
-            if visualize_graph:
-                plot_graph(selected_shapes, graph_edges)
             color_groups = color_graph(
-                num_nodes=int(len(to_color_shape_index)),
+                num_nodes=num_shapes,
                 graph_edge_indices=wp.array(graph_edges, dtype=wp.int32),
                 balance_colors=False,
             )
-            shape_color = np.zeros(model.shape_count, dtype=np.int32)
             num_colors = 0
             for group in color_groups:
                 num_colors += 1
                 shape_color[selected_shapes[group]] = num_colors
-        else:
-            # no edges in the graph, all shapes can collide with each other
-            shape_color = np.zeros(model.shape_count, dtype=np.int32)
+            if visualize_graph:
+                plot_graph(
+                    vertices=np.arange(num_shapes),
+                    edges=graph_edges,
+                    node_labels=[shape_keys[i] for i in selected_shapes] if shape_keys is not None else None,
+                    node_colors=[shape_color[i] for i in selected_shapes],
+                )
+
         return shape_color
 
     @override
@@ -1952,6 +1944,7 @@ class SolverMuJoCo(SolverBase):
         shape_size = model.shape_scale.numpy()
         shape_flags = model.shape_flags.numpy()
         shape_group = model.shape_group.numpy()
+        shape_mu = model.shape_material_mu.numpy()
 
         eq_constraint_type = model.equality_constraint_type.numpy()
         eq_constraint_body1 = model.equality_constraint_body1.numpy()
@@ -1988,14 +1981,6 @@ class SolverMuJoCo(SolverBase):
             GeoType.BOX: mujoco.mjtGeom.mjGEOM_BOX,
             GeoType.MESH: mujoco.mjtGeom.mjGEOM_MESH,
         }
-        geom_type_name = {
-            GeoType.SPHERE: "sphere",
-            GeoType.PLANE: "plane",
-            GeoType.CAPSULE: "capsule",
-            GeoType.CYLINDER: "cylinder",
-            GeoType.BOX: "box",
-            GeoType.MESH: "mesh",
-        }
 
         mj_bodies = [spec.worldbody]
         # mapping from Newton body id to MuJoCo body id
@@ -2025,15 +2010,15 @@ class SolverMuJoCo(SolverBase):
             selected_joints = np.where((joint_group == first_group) | (joint_group < 0))[0]
         else:
             # if we are not separating environments to worlds, we use all shapes, bodies, joints
-            selected_shapes = np.arange(model.shape_count)
-            selected_bodies = np.arange(model.body_count)
-            selected_joints = np.arange(model.joint_count)
+            selected_shapes = np.arange(model.shape_count, dtype=np.int32)
+            selected_bodies = np.arange(model.body_count, dtype=np.int32)
+            selected_joints = np.arange(model.joint_count, dtype=np.int32)
 
         # sort joints topologically depth-first since this is the order that will also be used
         # for placing bodies in the MuJoCo model
         joints_simple = list(zip(joint_parent[selected_joints], joint_child[selected_joints], strict=False))
         joint_order = topological_sort(joints_simple, use_dfs=True)
-        if any(joint_order != np.arange(len(joints_simple))):
+        if any(joint_order[i] != i for i in range(len(joints_simple))):
             warnings.warn(
                 "Joint order is not in depth-first topological order while converting Newton model to MuJoCo, this may lead to diverging kinematics between MuJoCo and Newton.",
                 stacklevel=2,
@@ -2043,19 +2028,19 @@ class SolverMuJoCo(SolverBase):
         # filter out shapes that are not colliding with anything
         colliding_shapes = selected_shapes[shape_flags[selected_shapes] & ShapeFlags.COLLIDE_SHAPES != 0]
 
-        # filter out non-colliding bodies using excludes
-        body_filters, shape_filters = self.find_body_collision_filter_pairs(
-            colliding_shapes,
-            self.model.shape_collision_filter_pairs,
-            self.model.shape_collision_group,
-            self.model.shape_body.numpy().tolist(),
-            self.model.body_shapes,
-        )
-
-        shape_color = self.color_collision_shapes(model, colliding_shapes)
-
         # number of shapes we are instantiating in MuJoCo (which will be replicated for the number of envs)
         colliding_shapes_per_env = len(colliding_shapes)
+
+        # filter out non-colliding bodies using excludes
+        body_filters = self.find_body_collision_filter_pairs(
+            model,
+            selected_bodies,
+            colliding_shapes,
+        )
+
+        shape_color = self.color_collision_shapes(
+            model, colliding_shapes, visualize_graph=False, shape_keys=model.shape_key
+        )
 
         # store selected shapes, bodies, joints for later use in update_geom_properties
         self.selected_shapes = wp.array(selected_shapes, dtype=wp.int32, device=model.device)
@@ -2075,7 +2060,7 @@ class SolverMuJoCo(SolverBase):
                 if skip_visual_only_geoms and not (shape_flags[shape] & ShapeFlags.COLLIDE_SHAPES):
                     continue
                 stype = shape_type[shape]
-                name = f"{geom_type_name[stype]}_{shape}"
+                name = f"{model.shape_key[shape]}_{shape}"
                 if stype == GeoType.PLANE and newton_body_id != -1:
                     raise ValueError("Planes can only be attached to static bodies")
                 geom_params = {
@@ -2125,17 +2110,13 @@ class SolverMuJoCo(SolverBase):
                         # collide with anything except shapes from the same color
                         geom_params["conaffinity"] = collision_mask_everything & ~contype
 
-                # use shape materials instead of defaults if available
-                if model.shape_material_mu is not None:
-                    shape_mu = model.shape_material_mu.numpy()
-                    if shape < len(shape_mu):
-                        # set friction from Newton shape materials using model's friction parameters
-                        mu = shape_mu[shape]
-                        geom_params["friction"] = [
-                            mu,
-                            model.rigid_contact_torsional_friction * mu,
-                            model.rigid_contact_rolling_friction * mu,
-                        ]
+                # set friction from Newton shape materials using model's friction parameters
+                mu = shape_mu[shape]
+                geom_params["friction"] = [
+                    mu,
+                    model.rigid_contact_torsional_friction * mu,
+                    model.rigid_contact_rolling_friction * mu,
+                ]
 
                 body.add_geom(**geom_params)
                 # store the geom name instead of assuming index
@@ -2376,6 +2357,8 @@ class SolverMuJoCo(SolverBase):
             "The number of geoms in the MuJoCo model does not match the number of colliding shapes in the Newton model."
         )
 
+        # add contact exclusions between bodies to ensure parent <> child collisions are ignored
+        # even when one of the bodies is static
         for b1, b2 in body_filters:
             mb1, mb2 = body_mapping[b1], body_mapping[b2]
             spec.add_exclude(bodyname1=spec.bodies[mb1].name, bodyname2=spec.bodies[mb2].name)