<!--
Thank you for contributing to Newton!

Please fill the relevant sections.

Checkboxes can also be marked after you submit the PR.
-->

## Description
Continuation of @camevor's work on https://github.com/newton-physics/newton/pull/793:
* Faster implementation of the body pairs to exclude in contact generation from SolverMuJoCo
* Only generate shape contact filters for colliding shapes
* Improvements to some graph plotting utility functions (only used for debugging)

## Newton Migration Guide

Please ensure the migration guide for **warp.sim** users is up-to-date with the changes made in this PR.

- [ ] The migration guide in ``docs/migration.rst`` is up-to date

## Before your PR is "Ready for review"

- [ ] Necessary tests have been added and new examples are tested (see `newton/tests/test_examples.py`)
- [ ] Documentation is up-to-date
- [ ] Code passes formatting and linting checks with `pre-commit run -a`