@eric-heiden 2025-09-20
The benchmark performance has decreased slightly, however I found this not to be severe since the overhead for computing those body pairs to exclude from collisions is rather small when comparing 8192 envs (3 runs each):

| Environment   | Before (ms)        | After (ms)         |
|---------------|--------------------|--------------------|
| Allegro hand  | 2233.59 ± 99.33    | 1908.22 ± 59.97    |
| G1 humanoid   | 2565.19 ± 215.68   | 2711.45 ± 133.25   |